update-draw ToDoList from App
